Docstrings

TerrariaWiringHelper.AndGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

TerrariaWiringHelper.CombGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

TerrariaWiringHelper.XorGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

TerrariaWiringHelper.fuse — Method

fuse(setup::Integer, input::Integer) = fuse(setup & input)

Return the state of a tile connected to multiple wires. Each bit of setup represents the state of a wire, and each bit of input represents the connection of a wire to the tile.

TerrariaWiringHelper.fuse — Method

fuse(input::Integer)

Return the state of a tile connected to multiple wires. Each bit of the input represents the state of a wire.

Examples

julia> TerrariaWiringHelper.fuse(0b0110)
false

julia> TerrariaWiringHelper.fuse(0b1011)
true

TerrariaWiringHelper.gate2string — Method

gate2string(G::CombGate)

Convert a CombGate into a readable string. See also CombGate, lampstate2string, logicsymb.

TerrariaWiringHelper.hasallcolor — Method

hasallcolor(G::AbstractGate, mask::Integer)
hasallcolor(mask::Integer) = Fix2(hasallcolor, mask)

Check if some gate is connect to some wires.

TerrariaWiringHelper.hascolor — Method

hascolor(lampstate::Integer, mask::Integer) = lampstate & mask
hascolor(mask::Integer) = Fix2(hascolor, mask)
hascolor(G::AbstractGate, mask::Integer) = any(hascolor(mask), lampstates(G))

Check if some lamp is connected to some wire (combination), or if some gate has a such lamp.

TerrariaWiringHelper.lampstate2string — Method

lampstate2string(lampstate::Integer, args::Integer)

Convert binary coded lampstate into a readable string. args tells the number of distinct inputs.

Examples

julia> lampstate2string(0b1111, 4)
"abcd"

julia> lampstate2string(0b11101, 4)
"~acd"

julia> lampstate2string(0b10000, 4)
"~"

TerrariaWiringHelper.logicsymb — Method

logicsymb(::typeof(AND|XOR)|::CombGate)

Return '&' or '^' according to input.