Docstrings

TerrariaWiringHelper.AndGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

TerrariaWiringHelper.CombGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

TerrariaWiringHelper.XorGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

TerrariaWiringHelper.Z2ColMat — Type

Z2ColMat{C<:Integer, R<:Integer}(data::Vector{C}, size::Int)
Z2ColMat(undef, m, n)
Z2ColMat(::AbstractMatrix)

$\mathbb{Z}_2$ matrix stored in a Vector where each entry represents a column. size specifies the height. C and R are used for dispatch; they specify the integer type used to represent a column vector and a row vector respectively. See also bit2type, Z2RowMat

Example

julia> Z2ColMat([1, 3, 5], 3)
3×3 Z2ColMat{UInt8, UInt8}:
 1  1  1
 0  1  0
 0  0  1

TerrariaWiringHelper.Z2RowMat — Type

Z2RowMat{C<:Integer, R<:Integer}(data::Vector{R}, size::Int)
Z2RowMat(undef, m, n)
Z2RowMat(::AbstractMatrix)

$\mathbb{Z}_2$ matrix stored in a Vector where each entry represents a row. size specifies the width. C and R are used for dispatch; they specify the integer type used to represent a column vector and a row vector respectively. See also bit2type, Z2ColMat

Example

julia> Z2RowMat([1, 3, 5], 3)
3×3 Z2RowMat{UInt8, UInt8}:
 1  0  0
 1  1  0
 1  0  1

TerrariaWiringHelper.Z2Vector — Type

Z2Vector{T<:Integer}(data::T, size::Int)
Z2Vector(::AbstractVector)

$\mathbb{Z}_2$ vector stored in an integer data where each bit represents an entry. size specifies the length.

Example

julia> Z2Vector(0b10110, 5)
5-element Z2Vector{UInt8}:
 0
 1
 1
 0
 1

TerrariaWiringHelper.CombLogic — Method

CombLogic(lamps, inputs, outputs::AbstractVector{Bool})

Return the lists of possible combinatory logics for given inputs and outputs.

Arguments

lamps::Integer: the number of logic lamps on the gate, from 2 to 7.

inputs::AbstractVector{<:Integer}: the list of input values specified by the lower bits of entries, with a 1 in front of the first bit.

outputs::AbstractVector{Bool}: the expected return values of the logic. This should have the same length as inputs, although it's not strictly checked.

Example

Imagine we want to check if a 4-bit input can be divided by 3. Further, the input is expected to have only one decimal digit, i.e. in 0:9. We don't care how the logic behaves for inputs 10:15. The truth table would be

in	out		in	out

0000	true		0101	false

0001	false		0110	true

0010	false		0111	false

0011	true		1000	false

0100	false		1001	true

Now we try to find a configuration with least lamps.

julia> inputs = 16:25; # `0:9` with the 5th last bit set to `1`

julia> outputs = [true, false, false, true, false, false, true, false, false, true];

julia> CombLogic(2, inputs, outputs) # there is no 2-lamp logic
Union{AndGate{Int64, StaticArraysCore.SVector{2, UInt8}}, XorGate{Int64, StaticArraysCore.SVector{2, UInt8}}}[]

julia> CombLogic(3, inputs, outputs) # there are 12 3-lamp logics, all of which use XOR gates
12-element Vector{Union{AndGate{Int64, StaticArraysCore.SVector{3, UInt8}}, XorGate{Int64, StaticArraysCore.SVector{3, UInt8}}}}:
 ^(~b, ad, c)
 ^(~b, abd, bc)
 ^(~ac, ad, ab)
 ^(~ac, abd, a)
 ^(~ac, cd, bc)
 ^(~ac, bcd, c)
 ^(~bd, c, a)
 ^(~bd, bc, ab)
 ^(~bd, cd, ad)
 ^(~bd, bcd, abd)
 ^(~acd, bc, c)
 ^(~acd, abd, ad)

See also CombGate

TerrariaWiringHelper.bit2type — Method

bit2type(n::Integer) :: DataType

Returns minimal integer type that contains n bits

Examples

julia> bit2type(5)
UInt8

julia> bit2type(9)
UInt16

julia> bit2type(32)
UInt32

julia> bit2type(33)
UInt64

julia> bit2type(100)
UInt128

TerrariaWiringHelper.datamask — Method

datamask(::Type{T}, size::Integer)
datamask(A::Z2Vector{T}) = datamask(T, A.size)

Internal. Returns the bitmask of type T where the lower size bits are set to 1.

Examples

julia> datamask(UInt8, 3)
0x07

julia> datamask(Z2Vector(0b10110, 5))
0x1f

julia> datamask(Z2Vector(0b10110, 4))
0x0f

TerrariaWiringHelper.fuse — Method

fuse(setup::Integer, input::Integer) = fuse(setup & input)

Return the state of a tile connected to multiple wires. Each bit of setup represents the state of a wire, and each bit of input represents the connection of a wire to the tile.

TerrariaWiringHelper.fuse — Method

fuse(input::Integer)

Return the state of a tile connected to multiple wires. Each bit of the input represents the state of a wire.

Examples

julia> TerrariaWiringHelper.fuse(0b0110)
false

julia> TerrariaWiringHelper.fuse(0b1011)
true

TerrariaWiringHelper.gate2string — Method

gate2string(G::CombGate)

Convert a CombGate into a readable string. See also CombGate, lampstate2string, logicsymb.

TerrariaWiringHelper.getbit — Method

getbit(x::Integer, n::Integer)

Get the n-th (counting from zero) last bit of x. See also setbit

Example

julia> [getbit(0b10110, n) for n in 0:4]
5-element Vector{Bool}:
 0
 1
 1
 0
 1

TerrariaWiringHelper.hasallcolor — Method

hasallcolor(G::AbstractGate, mask::Integer)
hasallcolor(mask::Integer) = Fix2(hasallcolor, mask)

Check if some gate is connect to some wires.

TerrariaWiringHelper.hascolor — Method

hascolor(lampstate::Integer, mask::Integer) = lampstate & mask
hascolor(mask::Integer) = Fix2(hascolor, mask)
hascolor(G::AbstractGate, mask::Integer) = any(hascolor(mask), lampstates(G))

Check if some lamp is connected to some wire (combination), or if some gate has a such lamp.

TerrariaWiringHelper.lampstate2string — Method

lampstate2string(lampstate::Integer, args::Integer)

Convert binary coded lampstate into a readable string. args tells the number of distinct inputs.

Examples

julia> lampstate2string(0b1111, 4)
"abcd"

julia> lampstate2string(0b11101, 4)
"~acd"

julia> lampstate2string(0b10000, 4)
"~"

TerrariaWiringHelper.logicsymb — Method

logicsymb(::typeof(AND|XOR)|::CombGate)

Return '&' or '^' according to input.

TerrariaWiringHelper.setbit — Method

setbit(x::Integer, y::Bool, n::Integer)

Set the n-th (counting from zero) last bit of x to y. See also getbit.

Example

julia> string(setbit(0b10110, true, 0), base = 2)
"10111"

julia> string(setbit(0b10110, false, 0), base = 2)
"10110"

TerrariaWiringHelper.type2bit — Method

type2bit(::Type{T})

The inverse of bit2type.

TerrariaWiringHelper.z2number — Method

z2number(::Number)
z2number(::AbstractMatrix{<:Number})

Convert Number to Bool in the $\mathbb{Z}_2$ sense.

Examples

julia> z2number(25)
true

julia> z2number(1:5)
5-element BitVector:
 1
 0
 1
 0
 1